Added naive proposal function

+++ src/ballot.rs
@@ -1,14 +1,14 @@
 #[derive(Debug, Clone)]
 pub struct Student {  // Perhaps rename because a double would count as 1 student
     pub name: String,
-    pub ballot: Vec<u8>,
+    pub ballot: Vec<f64>,
 }
 
 impl Student {
     pub fn new(name: String, num_houses: usize) -> Self {
         Self {
             name,
-            ballot: vec![0; num_houses]
+            ballot: vec![0.0; num_houses]
         }
     }
 }
